added convective fluxes to output
Noticed that freshwater budget was not closing in some cases. Think this was related to not taking account of changes in T/S due to the homogenise unstable layers routine (which we could think of as convection). So now the code outputs the fluxes associated with homogenisation, which does seem to close the updated budget code. Note that properly closing the budget also requires that you save out all of the timesteps (as opposed to a subset, which is more usual in the examples).